refactor: 線形漸化式を linear_recurrence.hpp に統合し kitamasa の NTT 版を削除
kitamasa の NTT 版は kth_term (Bostan-Mori) と機能重複かつ低速だったため削除し、
kitamasa は任意可換環版 O(k^2 log n) のみに絞った。kitamasa.hpp と bostan_mori.hpp を
lib/math/linear_recurrence.hpp に統合 (旧 2 ファイルは削除)。
- kitamasa: 畳み込み不能な環 (NTT 非対応 mod・dynamic_modint・行列・有理数) 用
- kth_term: NTT-friendly static_modint 用
