+++ test/aoj/alds1/kitamasa.test.cpp
@@ -1,8 +1,8 @@
 // competitive-verifier: PROBLEM https://onlinejudge.u-aizu.ac.jp/problems/ALDS1_10_A
-#include "math/kitamasa.hpp"
 #include <cstdint>
 #include <iostream>
 #include <vector>
+#include "math/linear_recurrence.hpp"
 
 int main(void) {
     int n;
